requirements.txt
PyMuPDF==1.24.11
sentence-transformers==3.2.0
faiss-cpu==1.9.0
transformers==4.45.2
torch==2.5.0
numpy==2.1.2
